scripts/tooling/blackhole_recovery/requirements.txt
pyocd
pyluwen @ git+https://github.com/tenstorrent/luwen.git@v0.6.5#subdirectory=crates/pyluwen
pyyaml
